refactor: move entrypoint for actors
