+++ src/module/foundryvtt-importer.ts
@@ -4,8 +4,8 @@ import { processInputJSON } from './journal';
 import { processTableJSON } from './table';
 import { renderSidebarButtons } from './renderSidebarButtons';
 import CONSTANTS from './constants';
-import { processActorInput } from './actor';
 import { processItemInput } from './item/input';
+import { processActorInput } from './actor/handleInput';
 
 Hooks.on('renderSidebarTab', (settings: Settings) => {
   if (!(game as Game)?.user?.isGM) return;
